+++ src/ska_low_mccs_spshw/daq_receiver/daq_receiver_device.py
@@ -4,7 +4,7 @@
 #
 # Distributed under the terms of the BSD 3-clause new license.
 # See LICENSE.txt for more info.
-
+# pylint: disable=too-many-lines
 """This module implements the MccsDaqReceiver device."""
 
 from __future__ import annotations  # allow forward references in type hints
@@ -249,6 +249,7 @@ class MccsDaqReceiver(SKABaseDevice):
         # `init_device` re-initialises any values defined in here.
         super().__init__(*args, **kwargs)
 
+        self.component_manager: DaqComponentManager
         self._health_state: HealthState = HealthState.UNKNOWN
         self._health_model: DaqHealthModel
         self._received_data_mode: str
@@ -350,6 +351,8 @@ class MccsDaqReceiver(SKABaseDevice):
 
         for command_name, method_name in [
             ("Stop", "stop_daq"),
+            ("StartDataRateMonitor", "start_data_rate_monitor"),
+            ("StopDataRateMonitor", "stop_data_rate_monitor"),
         ]:
             self.register_command_object(
                 command_name,
@@ -908,6 +911,37 @@ class MccsDaqReceiver(SKABaseDevice):
         (result_code, message) = handler()
         return ([result_code], [message])
 
+    @command(dtype_out="DevVarLongStringArray")
+    def StopDataRateMonitor(self: MccsDaqReceiver) -> DevVarLongStringArrayType:
+        """
+        Stop monitoring the data rate on the receiver interface.
+
+        :return: A tuple containing a return code and a string
+            message indicating status. The message is for
+            information purpose only.
+        """
+        handler = self.get_command_object("StopDataRateMonitor")
+        (result_code, message) = handler()
+        return ([result_code], [message])
+
+    @command(dtype_out="DevVarLongStringArray", dtype_in="DevLong")
+    def StartDataRateMonitor(
+        self: MccsDaqReceiver,
+        interval: float,
+    ) -> DevVarLongStringArrayType:
+        """
+        Start monitoring the data rate on the receiver interface.
+
+        :param interval: The interval in seconds at which to monitor the data rate.
+
+        :return: A tuple containing a return code and a string
+            message indicating status. The message is for
+            information purpose only.
+        """
+        handler = self.get_command_object("StartDataRateMonitor")
+        (result_code, message) = handler(interval)
+        return ([result_code], [message])
+
     @attribute(
         dtype=("str",),
         max_dim_x=2,  # Always the last result (unique_id, JSON-encoded result)
@@ -969,6 +1003,15 @@ class MccsDaqReceiver(SKABaseDevice):
         :return: the health report.
         """
         return self._health_model.health_report
+
+    @attribute(dtype="DevFloat")
+    def dataRate(self: MccsDaqReceiver) -> float | None:
+        """
+        Return the current data rate in Gb/s, or None if not being monitored.
+
+        :return: the current data rate in Gb/s, or None if not being monitored.
+        """
+        return self.component_manager.data_rate
 
 
 # ----------