+++ src/ska_low_mccs_spshw/tile/tile_device.py
@@ -3933,8 +3933,6 @@ class MccsTile(SKABaseDevice[TileComponentManager]):
             self._component_manager = component_manager
             super().__init__(logger)
 
-        SUCCEEDED_MESSAGE = "LoadCalibrationCoefficents command completed OK"
-
         def do(  # type: ignore[override]
             self: MccsTile.LoadCalibrationCoefficientsCommand,
             argin: list[float],
@@ -3977,10 +3975,10 @@ class MccsTile(SKABaseDevice[TileComponentManager]):
                 for i in range(1, len(argin), 8)
             ]
 
-            self._component_manager.load_calibration_coefficients(
+            result, message = self._component_manager.load_calibration_coefficients(
                 antenna, calibration_coefficients
             )
-            return (ResultCode.OK, self.SUCCEEDED_MESSAGE)
+            return (result, message)
 
     @command(dtype_in="DevVarDoubleArray", dtype_out="DevVarLongStringArray")
     def LoadCalibrationCoefficients(
@@ -4049,8 +4047,6 @@ class MccsTile(SKABaseDevice[TileComponentManager]):
             self._component_manager = component_manager
             super().__init__(logger)
 
-        SUCCEEDED_MESSAGE = "ApplyCalibration command completed OK"
-
         def do(
             self: MccsTile.ApplyCalibrationCommand,
             *args: Any,
@@ -4069,8 +4065,10 @@ class MccsTile(SKABaseDevice[TileComponentManager]):
             """
             switch_time = args[0]
 
-            self._component_manager.apply_calibration(switch_time)
-            return (ResultCode.OK, self.SUCCEEDED_MESSAGE)
+            return_code, message = self._component_manager.apply_calibration(
+                switch_time
+            )
+            return (return_code, message)
 
     @command(dtype_in="DevString", dtype_out="DevVarLongStringArray")
     def ApplyCalibration(self: MccsTile, argin: str) -> DevVarLongStringArrayType: